refactor(contracts): decompose DocumentsService into Template/Signature/Rendering/Persistence + format util
Behavior-preserving (1 $tx doc+audit metadata, 0 JE). Facade keeps 20-method surface +
4-arg ctor + delegates; internally constructs 4 sub-services (3 specs + module + 2
controllers untouched). The ensureSignedContractDocument $tx moved WHOLE; the
fire-and-forget signContract->ensureSignedContractDocument hook kept non-awaited/non-tx;
replacePlaceholders (~445 LOC, load-bearing pass ordering) + default-template HTML moved
VERBATIM. tsc 0, contracts doc/template/signing tests green, lint 0.
